refactor: split UIPreferencesTab into focused sub-components
Part of #103
